requirements.txt
# str #
six                 # Python 2 and 3 compatibility utilities
chardet             # Universal encoding detector for Python 2 and 3
python-levenshtein  # Python extension for computing string edit distances and similarities.
